move functionality around, clean up controller method
