+++ src/main/java/util/SessionUtils.java
@@ -22,27 +22,6 @@ public class SessionUtils {
 
     private static final Log log = LogFactory.getLog(SessionUtils.class);
 
-    public static SerializableFormSession loadSessionFromJson(String sessionJSON) {
-        SerializableFormSession session = new SerializableFormSession();
-        JSONObject sessionObject = new JSONObject(sessionJSON);
-        JSONObject sessionData = sessionObject.getJSONObject("session_data");
-        session.setTitle(sessionData.getString("session_name"));
-        session.setUsername(sessionData.getString("username"));
-        session.setDomain(sessionData.getString("domain"));
-        session.setOneQuestionPerScreen(false);
-        session.setAsUser(null);
-        session.setInstanceXml(sessionObject.getString("instance"));
-        session.setFormXml(sessionObject.getString("xform"));
-        session.setInitLang(sessionObject.getString("init_lang"));
-        session.setSequenceId(sessionObject.getInt("seq_id"));
-        HashMap<String, String> sessionDataMap = new HashMap<>();
-        session.setSessionData(sessionDataMap);
-        if (sessionData.has("case_id")) {
-            sessionDataMap.put("case_id", (String) sessionData.get("case_id"));
-        }
-        return session;
-    }
-
     public static String getBestTitle(SessionWrapper session) {
 
         String[] stepTitles;
